@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 0.1.3
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: '>=3.5.2 <4.0.0'
 
 dependencies:
   cupertino_icons: ^1.0.2
@@ -45,7 +45,7 @@ dependencies:
   flutter:
     sdk: flutter
   bloc: ^8.1.4
-
+    
 
   injectable: ^2.4.2
   equatable: ^2.0.5
@@ -70,7 +70,7 @@ dependencies:
   shimmer: ^3.0.0
   showcaseview: ^3.0.0
   flutter_slidable: ^3.1.1
-
+  flutter_contacts: ^1.1.9+2
 dev_dependencies:
 
   flutter_test:
