@@ -2,7 +2,7 @@ import 'package:bulk_app/core/theming/colors.dart';
 import 'package:bulk_app/core/theming/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter_svg/svg.dart';
+import 'package:vector_graphics/vector_graphics_compat.dart';
 
 class AccountInformation extends StatelessWidget {
   const AccountInformation(
@@ -25,7 +25,8 @@ class AccountInformation extends StatelessWidget {
         padding: EdgeInsets.symmetric(horizontal: 18.sp),
         child: Row(
           children: [
-            SvgPicture.asset(imagePath),
+            
+            VectorGraphic(loader: AssetBytesLoader(imagePath)),
             SizedBox(width: 12.w),
             Text(
               "Contacts",
